+++ src/_h5ai/client/js/inc/core/resource.js
@@ -32,37 +32,31 @@ modulejs.define('core/resource', ['_', '$', 'config', 'core/settings'], function
 			}
 
 			return fallbackHref + 'file.svg';
-		},
-
-		loadScript = function (filename, callback) {
-
-			$.ajax({
-					url: scriptsHref + filename,
-					dataType: 'script'
-				}).done(function () {
-
-					callback();
-				});
-		},
-
-		loadScriptGlob = function (filename, globalId, callback) {
-
-			if (win[globalId]) {
-				callback(win[globalId]);
-			} else {
-				loadScript(filename, function () { callback(win[globalId]); });
-			}
-		},
-
-		ensureSH = function (callback) {
-
-			loadScriptGlob('syntaxhighlighter.js', 'SyntaxHighlighter', callback);
 		};
+
+		// loadScript = function (filename, callback) {
+
+		// 	$.ajax({
+		// 			url: scriptsHref + filename,
+		// 			dataType: 'script'
+		// 		}).done(function () {
+
+		// 			callback();
+		// 		});
+		// },
+
+		// loadScriptGlob = function (filename, globalId, callback) {
+
+		// 	if (win[globalId]) {
+		// 		callback(win[globalId]);
+		// 	} else {
+		// 		loadScript(filename, function () { callback(win[globalId]); });
+		// 	}
+		// };
 
 
 	return {
 		image: image,
-		icon: icon,
-		ensureSH: ensureSH
+		icon: icon
 	};
 });